+++ internal/sdkprovider/service/serviceintegration/service_integration_endpoint.go
@@ -16,6 +16,21 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
+var integrationEndpointTypes = []string{
+	"datadog",
+	"prometheus",
+	"rsyslog",
+	"external_elasticsearch_logs",
+	"external_opensearch_logs",
+	"external_aws_cloudwatch_logs",
+	"external_google_cloud_logging",
+	"external_kafka",
+	"jolokia",
+	"signalfx",
+	"external_schema_registry",
+	"external_aws_cloudwatch_metrics",
+}
+
 var aivenServiceIntegrationEndpointSchema = map[string]*schema.Schema{
 	"project": {
 		Description: "Project the service integration endpoint belongs to",
@@ -30,10 +45,12 @@ var aivenServiceIntegrationEndpointSchema = map[string]*schema.Schema{
 		Type:        schema.TypeString,
 	},
 	"endpoint_type": {
-		Description: "Type of the service integration endpoint",
-		ForceNew:    true,
-		Required:    true,
-		Type:        schema.TypeString,
+		Description: "Type of the service integration endpoint. Possible values: " +
+			schemautil.JoinQuoted(integrationEndpointTypes, ", ", "`"),
+		ForceNew:         true,
+		Required:         true,
+		Type:             schema.TypeString,
+		ValidateDiagFunc: schemautil.ValidateEnum(integrationEndpointTypes...),
 	},
 	"endpoint_config": {
 		Description: "Integration endpoint specific backend configuration",