+++ examples/resources/aiven_clickhouse_user/import.sh
@@ -1,1 +1,1 @@
-terraform import aiven_clickhouse_user.ch-user project/service_name/id
+terraform import aiven_clickhouse_user.ch-user project/service_name/username