+++ examples/resources/aiven_m3db_user/import.sh
@@ -1,0 +1,1 @@
+terraform import aiven_m3db_user.foo project/service_name/username