+++ CHANGELOG.md
@@ -8,6 +8,8 @@ nav_order: 1
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Added docs and validation for `aiven_service_integration_endpoint`
+
 ## [4.2.1] - 2023-04-06
 
 - Add GCP virtual network peering example