+++ CHANGELOG.md
@@ -14,6 +14,7 @@ nav_order: 1
 - Fix Account SAML Field mapping set method
 - Adjust generated SQL for ClickHouse privilege grants
 - Fix `required` not generated for top level fields for user config options
+- Fix `required` properties not always sent to the API for user config options
 
 ## [4.2.1] - 2023-04-06