+++ CHANGELOG.md
@@ -15,6 +15,7 @@ nav_order: 1
 - Adjust generated SQL for ClickHouse privilege grants
 - Fix `required` not generated for top level fields for user config options
 - Deprecate `karapace` option on `aiven_kafka` schema
+- Deprecate service `disk_space` field
 
 ## [4.2.1] - 2023-04-06